+++ awc/src/client/pool.rs
@@ -19,7 +19,7 @@ use actix_rt::time::{sleep, Sleep};
 use actix_service::Service;
 use ahash::AHashMap;
 use futures_core::future::LocalBoxFuture;
-use futures_util::FutureExt;
+use futures_util::FutureExt as _;
 use http::uri::Authority;
 use pin_project_lite::pin_project;
 use tokio::sync::{OwnedSemaphorePermit, Semaphore};